+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## vNext
 
-- Add custom `InstrumentationScope` support via
+- Add custom instrumentation scope attributes support via
   `OpenTelemetryTracingBridge::builder_with_scope_attributes(..)`.
   [3415](https://github.com/open-telemetry/opentelemetry-rust/issues/3415)